add check for packer-builder-type
Check if packer-builder-type is vmware-iso we don't delete linux
headers, gcc and ensure that the vmware-tools kernel modules will be
automaticly rebuildet on kernel update.

+++ scripts/ubuntu/cleanup.sh
@@ -1,10 +1,16 @@
 #!/bin/bash -eux
 
 # delete all linux headers
-dpkg --list | awk '{ print $2 }' | grep linux-headers | xargs apt-get -y purge
+if [[ "$PACKER_BUILDER_TYPE" == "vmware-iso" ]]; then
+  sed -i.bak 's/answer AUTO_KMODS_ENABLED_ANSWER no/answer AUTO_KMODS_ENABLED_ANSWER yes/g' /etc/vmware-tools/locations
+  sed -i 's/answer AUTO_KMODS_ENABLED no/answer AUTO_KMODS_ENABLED yes/g' /etc/vmware-tools/locations
+  dpkg --list | awk '{ print $2 }' | grep linux-headers | grep -v `uname -r` | xargs apt-get -y purge
+else
+  dpkg --list | awk '{ print $2 }' | grep linux-headers | xargs apt-get -y purge
+fi
 
 # this removes specific linux kernels, such as
-# linux-image-3.11.0-15-generic but 
+# linux-image-3.11.0-15-generic but
 # * keeps the current kernel
 # * does not touch the virtual packages, e.g.'linux-image-generic', etc.
 #
@@ -17,7 +23,9 @@ dpkg --list | awk '{ print $2 }' | grep linux-source | xargs apt-get -y purge
 dpkg --list | awk '{ print $2 }' | grep -- '-dev$' | xargs apt-get -y purge
 
 # delete compilers and other development tools
-apt-get -y purge cpp gcc g++
+if [[ "$PACKER_BUILDER_TYPE" != "vmware-iso" ]]; then
+  apt-get -y purge cpp gcc g++
+fi
 
 # delete X11 libraries
 apt-get -y purge libx11-data xauth libxmuu1 libxcb1 libx11-6 libxext6
